requirements.txt
PyJWT==2.8.0
flake8==6.1.0 
flask==3.0.0
mysql-connector-python==8.2.0 
pytest==7.4.3 
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1 
mysqlclient==2.2.0
yagmail==0.15.293
celery==5.3.5 
SQLAlchemy==2.0.23
mysqlclient==2.2.0
alembic==1.12.1 